requirements.txt
importlib-metadata==4.13.0
kf_lib_data_ingest @ git+https://github.com/kids-first/kf-lib-data-ingest.git@no-retry-session
kf_utils @ git+https://github.com/kids-first/kf-utils-python.git@no-retry-session
pandas<1.3,>=1.1
python-dotenv
SQLAlchemy==1.3.20
